requirements.txt
amqp 
asgiref 
async-timeout==4.0.2
backports.functools-lru-cache 
beautifulsoup4
billiard 
brotlipy==0.7.0
celery 
certifi==2022.5.18.1
cffi 
charset-normalizer==2.0.12
click 
click-didyoumean 
click-plugins==1.1.1
click-repl 
colorama 
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django 
django-cors-headers 
django-filter 
django-templated-mail==1.1.1
djangorestframework 
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna 
importlib-metadata 
itypes==1.2.0
Jinja2==3.1.2
kombu 
konlpy
Markdown 
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
prompt-toolkit 
psycopg2 
pycparser 
PyJWT==2.4.0
pyOpenSSL 
pyparsing==3.0.9
PySocks 
python3-openid==3.2.0
pytz==2021.3
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
six 
social-auth-app-django==4.0.0
social-auth-core==4.3.0
spacy
spacy-udpipe
spacy-thai
spacy-stanza
suparkanbun
sqlparse 
stripe 
typing_extensions 
uritemplate==4.1.1
urllib3
unidic2ud
underthesea 
vine 
wcwidth 
win-inet-pton 
wincertstore==0.2
wrapt==1.14.1
zipp 
